refactor(dialogue_manager): extract retrieval/summary/learner_update modules
Step 4/9. Move three small impl blocks out of mod.rs into focused
sibling modules:
- retrieval.rs: retrieve_knowledge, retrieve_long_term_memory
- summary.rs: refresh_summary_if_due, refresh_summary_if_stale,
  regenerate_summary_through
- learner_update.rs: update_learner_model (the placeholder
  word-count engagement heuristic)

Visibility tightens to pub(super) on the cross-module entry points.
regenerate_summary_through stays private to summary.rs since both
callers (refresh_*) are co-located. Behavior unchanged; same
constants, same magic numbers, same docstrings.

mod.rs: 2904 → 2776 lines.
